api payment and update status

+++ src/main/java/com/example/api_movie/dto/FoodDto.java
@@ -5,6 +5,8 @@ import jakarta.persistence.EnumType;
 import jakarta.persistence.Enumerated;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Getter
 @Setter
 @AllArgsConstructor
@@ -14,7 +16,7 @@ public class FoodDto {
     @Enumerated(EnumType.STRING)
     private Food.Category category;
     private String name;
-    private int price;
+    private BigDecimal price;
     private String detail;
     private String img;
 }
